requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
billiard==4.2.1
black==24.10.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.9
cron-descriptor==1.4.5
Django==5.1.4
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
flake8==7.1.1
idna==3.10
inflection==0.5.1
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.2.0
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
pyparsing==3.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.5.0
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
